requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
email-validator
passlib
python-jose
